Get those null checks out of my sight
This properly fixes an iterator invalidation bug that was causing spurious crashes.

When an object deletes something as a result of their onCollision() handler,
we *cannot* also delete it from collisionPairs,
as that would invalidate the iterator when we get to that point.

This was patched by null checks everywhere, but since we were using an invalid iterator at that point,
the null checks would only stop us from crashing but we wouldn't be unable to finish looping through all the pairs.

Now, we keep track of the deleted-but-still-in-collisionPairs objects, and skip over those when iterating.
After we finish our loop, we erase all of them from collisionPairs and clear the vector to start again.