Fix the game over menu!
The health bar was getting scaled to 0 (as expected).
When we try to reverse the transformation though,
we end up dividing by zero and corrupting the transform.

Thus, rather than manually reversing, just copy the AT
and re-assign it when we need to undo all the transformations.

Need to tell Floryan this.